[OSD-17714] Updated sts pkg
